Move SSE parser from perplexity.rs to llm.rs
