+++ wkfl/src/llm/perplexity.rs
@@ -1,13 +1,13 @@
-use std::{
-    io::{BufRead, BufReader, Read},
-    str::FromStr,
-};
+use std::str::FromStr;
 
 use anyhow::anyhow;
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::config::{resolve_secret, Config};
+use crate::{
+    config::{resolve_secret, Config},
+    llm::SseReader,
+};
 
 #[derive(Debug, Default, Serialize, Deserialize)]
 #[serde(rename_all = "kebab-case")]
@@ -69,58 +69,6 @@ pub struct PerplexityResponse {
     pub citations: Option<Vec<String>>,
 }
 
-pub struct ServerSentEvent {
-    pub data: String,
-    pub is_done: bool,
-}
-
-pub struct SseReader {
-    reader: BufReader<Box<dyn Read>>,
-}
-
-impl SseReader {
-    pub fn new(reader: Box<dyn Read>) -> Self {
-        Self {
-            reader: BufReader::new(reader),
-        }
-    }
-
-    pub fn next_event(&mut self) -> anyhow::Result<Option<ServerSentEvent>> {
-        let mut line = String::new();
-        loop {
-            if let Err(e) = self.reader.read_line(&mut line) {
-                return Err(anyhow!(e));
-            }
-            
-            if line.trim().is_empty() {
-                line.clear();
-                continue;
-            }
-
-            // Parse data prefix
-            let data_str = match line.strip_prefix("data: ") {
-                Some(data_str) => data_str,
-                None => {
-                    return Err(anyhow!("Invalid data prefix '{}'", line));
-                }
-            };
-
-            // Check if stream is done
-            if data_str.starts_with("[Done]") {
-                return Ok(Some(ServerSentEvent {
-                    data: String::new(),
-                    is_done: true,
-                }));
-            }
-
-            return Ok(Some(ServerSentEvent {
-                data: data_str.to_string(),
-                is_done: false,
-            }));
-        }
-    }
-}
-
 pub struct PerplexityStreamResponseIterator {
     sse_reader: SseReader,
     done: bool,
@@ -133,20 +81,20 @@ impl Iterator for PerplexityStreamResponseIterator {
         if self.done {
             return None;
         }
-        
+
         // Get the next SSE event
         let event = match self.sse_reader.next_event() {
             Ok(Some(event)) => event,
             Ok(None) => return None,
             Err(e) => return Some(Err(e)),
         };
-        
+
         // Handle done event
         if event.is_done {
             println!("Stream completed");
             return None;
         }
-        
+
         // Parse the response
         let response = match serde_json::from_str::<PerplexityResponse>(&event.data) {
             Ok(response) => response,
